chore: Group financial commitment report tools.
