+++ src/tools/financial-commitment-reports/index.ts
@@ -1,0 +1,1 @@
+import "./list-financial-commitment-reports";
