+++ src/tools/index.ts
@@ -8,6 +8,7 @@ import "./create-folder";
 import "./create-virtual-tag-config";
 import "./delete-cost-report";
 import "./delete-folder";
+import "./financial-commitment-reports";
 import "./get-anomaly";
 import "./get-cost-alert";
 import "./get-cost-provider-accounts";
@@ -31,7 +32,6 @@ import "./list-cost-reports";
 import "./list-cost-services";
 import "./list-costs";
 import "./list-dashboards";
-import "./list-financial-commitment-reports";
 import "./list-folders";
 import "./list-provider-resources";
 import "./list-recommendations";
